Fix: stdError

+++ libft/ft_atoll.c
@@ -6,7 +6,7 @@
 /*   By: hekang <hekang@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/02 22:40:21 by hekang            #+#    #+#             */
-/*   Updated: 2021/06/22 18:58:04 by hekang           ###   ########.fr       */
+/*   Updated: 2021/06/24 09:38:05 by hekang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 long long				ft_atoll(const char *str)
 {
 	int					cnt;
-	unsigned long int	result;
+	long long			result;
 	int					minus;
 
 	cnt = 0;
